# Vectorize Guid.ToString