@EgorBo 2023-02-05
@stephentoub @MihaZupan anything else? 
I'll simplify this impl even further by sharing code with HexConverter (and, probably, Utf8Formatter.Guid) and moving hand-made Vector128 crossplat helpers to Vector128 itself (clean up) in my next PR